+++ main/feature/src/boofcv/abst/feature/describe/WrapDescribeBrief.java
@@ -54,7 +54,7 @@ public class WrapDescribeBrief<T extends ImageBase> implements ExtractFeatureDes
 	}
 
 	@Override
-	public TupleDesc_F64 process(int x, int y, double orientation, double scale, TupleDesc_F64 ret) {
+	public TupleDesc_F64 process(double x, double y, double orientation, double scale, TupleDesc_F64 ret) {
 
 		if( !alg.process(x,y,feature) )
 			return null;
